Move experimental features to its own package and refactor
