+++ WordPress/src/main/AndroidManifest.xml
@@ -236,7 +236,7 @@
             android:label="@string/me_btn_app_settings"
             android:theme="@style/WordPress.NoActionBar" />
         <activity
-            android:name=".ui.prefs.ExperimentalFeaturesActivity"
+            android:name=".ui.prefs.experimentalfeatures.ExperimentalFeaturesActivity"
             android:label="@string/me_btn_experimental_features"
             android:theme="@style/WordPress.NoActionBar" />
         <activity
